WIP: DArray and NOf.
DArray is a parser for a dynamic array of elements, where there is a
static maximum capacity but dynamic size, read from a subparser.
NOf is essentially the same specialized to have unit return value and
elements of unit type; may be worth just enabling the feature and
specializing DArray instead.

Currently, the DArray implementation is not adequately constraining the
length field to be a usize.