Fix pair parser, add LengthFallback.
LengthFallback allows you specify a sub-parser for a field that is
bounded by a length, and effectively catch a failed parse of the
embedded data.
ObserveLengthedBytes combines the sub-parser interpretation with
ObserveBytes; the observer will observe the whole length of the
LengthFallback except for it's initial length field, allowing you to
hash a transaction where the subparser still rejected.