Reduce polymorphism in json structs, and duplication.
define_json_struct! and define_json_struct_interp! now define separate
new types for both the JsonInterp struct and the Schema type, rather
than using type synonyms for particular instances of the generic
container. The generic container type persists as the return value,
given that we don't have much constraint on the types it will need to
contain.

This also enables us to write the trait constraints on the Interp struct
itself, which should reduce the scope we see type errors at and improve
the working experience with the library.

define_json_struct! also now defines a macro that will call
define_json_struct_interp for you with the original definition, which
preserves the separation of schema and interpretation but removes the
code duplication.