Add proto-gen
A very simple setup that uses the protobuf package to generate
types to parse the data comming from protoc. Then it uses protoc
to parse tendermint/abci/types.proto and generate a file containing
FileDescriptorSet data for the proto file. The script then uses
the protobuf package to decode that file into a FileDescriptorSet
struct and then it prints that struct out.